% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_clmm.R
\name{pool.clmm}
\alias{pool.clmm}
\title{Calculate pooled estimates of fixed effects, standard deviation of the
random effect and conditional modes of the random effect terms.}
\usage{
pool.clmm(x)
}
\arguments{
\item{x}{an mira object returned from the with.mids function from the mice
package.}
}
\value{
A clmm.pool object.
}
\description{
Calculate pooled estimates of fixed effects, standard deviation of the
random effect and conditional modes of the random effect terms.
}
