% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_clmm.R
\name{pooling.clmm}
\alias{pooling.clmm}
\title{Calculate pooled estimates of fixed effects, standard deviation of the
random effect and conditional modes of the random effect terms.}
\usage{
\method{pooling}{clmm}(x, conf.int.re = c("none", "profile"), data = NULL)
}
\arguments{
\item{x}{A list containing clmm fits. For instance the analyses slot from
a mira object of the mice package. Or the result from the with.mids
function from the mice package.}

\item{conf.int.re}{Type of 95% confidence interval for the variance of the
random effect to be calculated. Default is none.}

\item{data}{mira object used to fit the model. Needed for calculating the profile likelihood.}
}
\value{
A pooled.clmm object. With the following elements:
}
\description{
Calculate pooled estimates of fixed effects, standard deviation of the
random effect and conditional modes of the random effect terms.
}
\section{Slots}{

\describe{
\item{\code{fixed_effects}}{Pooled fixed effect estimates and associated standard errors.}

\item{\code{random_effects}}{Pooled modes of random effects with associated conditional variances.}

\item{\code{random_dist}}{Standard deviation of the random effect distribution and associated median odds ratio}

\item{\code{conf_in_re}}{Optional 95% confidence interval of the variance of the random effects.}
}}

