% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.clmm.r
\name{prof_ci}
\alias{prof_ci}
\title{Compute the profiled likelihood for the standard deviation for the random
effect in a fitted cumulative mixed model.}
\usage{
prof_ci(fits, alpha = 0.05, index, data)
}
\description{
Compute the profiled likelihood for the standard deviation for the random
effect in a fitted cumulative mixed model.
}
