% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pooled.clmm.R
\name{plot.pooled.clmm}
\alias{plot.pooled.clmm}
\title{Function to visualize the effect of the clusters on the outcome}
\usage{
\method{plot}{pooled.clmm}(x, ...)
}
\arguments{
\item{x}{a clmm.pool object returned from the pool.clmm function.}
}
\value{
A plot showing the conditional modes of the clusters with the 95%
prediction intervals.
}
\description{
Function to visualize the effect of the clusters on the outcome
}
